requirements.txt
tensorflow-cpu==2.5.0
matplotlib==3.4.2